Remove Set and Unit and rename/remove tests.
